Simplify subprocess spawning
Previously, the command was first executed once, and if that resulted in
a certain WindowsError, then the command was retried in shell mode. The
intention was to support Windows' file associations for file launch.
However, this causes problems in non-Windows environments, as
WindowsError exception is not defined.

Trying to catch a WindowsError exception is unnecessary for file
associations, as non-executable files are already run in shell mode on
the first try. This change simplifies the logic, still supporting direct
execution for executables and shell mode for shell builtins and Window's
file launch.

Fixes #18.