Warn users about using functions directly imported from package
Vistir allows doing from `vistir import foo` where `foo` is
actually in module `vistir.bar`.

Which means that projects that vendor vistir (like pipenv) must ship
also modules which are not used. That means that they must be shipping
`bar.py` just for the `import foo from bar` inside `vistir/__init__.py`
to succeed. With this change, we allow for 3rd party to remove the
module bar if it isn't used.